pyproject.toml
[project]
name = "github-commit-summarizer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.0.0",
    "requests>=2.32.5",
    "groq>=0.4.0",
    "google-generativeai>=0.3.0",
    "dspy-ai>=2.4.0",
    "openai>=1.0.0",
]
